create empty block
